refactor(BreadcrumbMenu): remove the interface to index file
