--- src/components/navigation/BreadcrumbMenu/interfaces/BreadcrumbMenu.interface.ts
@@ -1,5 +1,0 @@
-import { IBreadcrumbRoute } from "./BreadcrumbMenu.route.interface";
-
-export interface IBreadcrumbMenuProps {
-  routes: IBreadcrumbRoute[];
-}
