--- src/components/navigation/BreadcrumbMenu/interfaces/BreadcrumbMenu.route.interface.ts
@@ -1,5 +1,0 @@
-export interface IBreadcrumbRoute {
-  id: string;
-  path: string;
-  label: string;
-}
